Extract submission endpoint to separate file
